Update Caddy dependency
As of Caddy v2.0.0-beta.20, the split_path config parameter takes a slice/array. This commit adapt to the change. Through adapting to the update, a bug in processing `fastcgi_split_path_info` was found. The `fastcgi_split_path_info` directive in NGINX takes a regexp with 2 capture groups: the first is for the CGI script file name including its extension, and the second is for PATH_INFO. Before this commit, the directive was treated as if it takes the extension of the script file. The fix in this commit finds the extension by looking for a period followed by a series of alphanumeric characters, and uses that as our split_path paramter.